Update pooling options for v3 protocol
Increase the number of parallel open connections used by v3 and align it
with the defaults for v2.

While v3 allows 32K concurrent operations vs v2 that allows only 256
concurrent operations the usage of a single connection does not allow to
scale to the maximal performance. Allowing multiple connections (as v2
allows) provides better performance.

This is esepcially true in cases in which the client is actually able to
utilize this (https://issues.apache.org/jira/browse/CASSANDRA-9558) to
that end cassandra-stress has been updated to be forced to use v2 which
allows multiple parallel connections instead of v3.

Signed-off-by: Shlomi Livne <shlomi@scylladb.com>